backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
Django==4.2
djangorestframework==3.15.2
drf-yasg==1.21.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
openai==1.57.3
packaging==24.2
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
